Completely reworked the Vigenere algorithm, works now.
Mind that this version only creates a key that fits into the length of the given text.
It will just copy the key over and over again until it has the same length of the text.

There is another version actually which will just attach all of the other signs which are not in use for the key.
Anyway, I decided not to do that as this is very dependent on the system and used encoding table.